+++ cryptography/src/ed25519/core/batch.rs
