+++ cryptography/src/ed25519/core/error.rs
