+++ cryptography/src/ed25519/core/signature.rs
