+++ cryptography/src/ed25519/core/signing_key.rs
