+++ cryptography/src/ed25519/core/verification_key.rs
